requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
letta-client==0.1.256
markdown-it-py==4.0.0
mdurl==0.1.2
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
python-dotenv==1.1.1
rich==14.1.0
sniffio==1.3.1
typing-extensions==4.14.1
typing-inspection==0.4.1
